made activity into fragment
