+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.1Direction">
-        <activity android:name=".eventviewer.DisplayEvent"></activity>
+        <activity android:name=".navigation.fragment.home.DisplayEvent"></activity>
         <activity
             android:name=".navigation.NavigationActivity"
             android:label="@string/title_activity_navigation"
@@ -26,7 +26,7 @@
 
         </activity> <!-- Why on god's green earth are those lines needed for EventCreatorTest to pass ? -->
         <activity android:name=".events.ui.EventCreator" />
-        <activity android:name=".eventviewer.EventView" />
+        <activity android:name=".navigation.fragment.home.EventView" />
 
         <receiver android:name=".notification.NotificationPublisher" />
