--- app/src/main/java/com/github/onedirection/navigation/fragment/home/HomeViewModel.java
@@ -1,19 +1,0 @@
-package com.github.onedirection.navigation.fragment.home;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class HomeViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public HomeViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is an example text for Home");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
