+++ app/src/main/res/layout/activity_display_event.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:onClick="buttonStartEditEvent"
-    tools:context=".eventviewer.DisplayEvent">
+    tools:context=".navigation.fragment.home.DisplayEvent">
 
     <TextView
         android:id="@+id/eventEndTimeDisplay"
