+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -5,14 +5,17 @@
     tools:showIn="navigation_view">
 
     <group android:checkableBehavior="single">
+
         <item
             android:id="@+id/nav_home"
             android:icon="@drawable/ic_menu_home"
             android:title="@string/menu_home" />
+
         <item
             android:id="@+id/nav_calendar"
             android:icon="@drawable/ic_menu_calendar"
             android:title="@string/menu_calendar" />
+
         <item
             android:id="@+id/nav_map"
             android:icon="@drawable/ic_menu_map"
@@ -34,9 +37,6 @@
             android:icon="@drawable/ic_menu_logout"
             android:visible="false"
             android:title="@string/menu_logout" />
-        <item
-            android:id="@+id/nav_event_viewer"
-            android:icon="@drawable/ic_menu_view"
-            android:title="@string/menu_event_viewer"/>
+
     </group>
 </menu>
